refactor(solver): split conditional infer helpers
